convert back to extension and bypass policy violation error by importing rpack during runtime
