+++ __init__.py
@@ -1,15 +1,16 @@
 #The init file for the plugin
-bl_info = {
-    "name" : "KKBP (Koikatsu Blender Porter)",
-    "author" : "a blendlet and some blenderchads",
-    "location" : "View 3D > Tool Shelf > KKBP and Image Editor > Tool Shelf > KKBP",
-    "description" : "Scripts to automate cleanup of a Koikatsu export",
-    "version": (7, 0, 0),
-    "blender" : (4, 2, 0),
-    "category" : "3D View",
-    "tracker_url" : "https://github.com/FlailingFog/KK-Blender-Porter-Pack/",
-    "doc_url": "https://github.com/FlailingFog/KK-Blender-Porter-Pack/blob/master/wiki/Wiki%20top.md",
-}
+# bl_info = {
+#     "name" : "KKBP (Koikatsu Blender Porter)",
+#     "author" : "a blendlet and some blenderchads",
+#     "location" : "View 3D > Tool Shelf > KKBP and Image Editor > Tool Shelf > KKBP",
+#     "description" : "Scripts to automate cleanup of a Koikatsu export",
+#     "version": (7, 0, 0),
+#     "blender" : (4, 2, 0),
+#     "category" : "3D View",
+#     "tracker_url" : "https://github.com/FlailingFog/KK-Blender-Porter-Pack/",
+#     "doc_url": "https://github.com/FlailingFog/KK-Blender-Porter-Pack/blob/master/wiki/Wiki%20top.md",
+# }
+
 
 from bpy.utils import register_class, unregister_class
 from bpy.types import Scene
