+++ rpack/_core.pyx
